move to ocr methods class
