+++ BarcodeScanner.Mobile.Maui/Platforms/Android/CameraViewHandler.android.cs
@@ -70,14 +70,8 @@ namespace BarcodeScanner.Mobile
                                 .SetTargetResolution(TargetResolution())
                                 .Build();
 
-            if (VirtualView.IsOCR)
-            {
-                imageAnalyzer.SetAnalyzer(_cameraExecutor, new OCRAnalyzer(VirtualView));
-            }
-            else
-            {
-                imageAnalyzer.SetAnalyzer(_cameraExecutor, new BarcodeAnalyzer(VirtualView));
-            }
+
+            imageAnalyzer.SetAnalyzer(_cameraExecutor, new BarcodeAnalyzer(VirtualView));
 
             var cameraSelector = SelectCamera(cameraProvider);
